Implement the storage backend for the project application data.
The project application data is stored in an updatable view. Access is
handled by setting users variables which contain security tokens which
are only available after authorization, that is: either a user is logged
into the cloud or we have a special authenticated public share with
password protection. Here the token is the user's email address.

A users can access the application data-set by either logging in or
aquiring a temporary password for the share.

Working: passwords can be aquired by email.

Missing: notification about the created share by email.

This should be the next step, the registration token and a summary of
the submitted values have to be send out to the user.